Refactoring - PN-221
- 밝기조절
